Move all constant classes out of blocks
